requirements.txt
# Network Automation Dependencies
netmiko>=4.2.0
paramiko>=3.4.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
cryptography>=41.0.7
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.7
markdown>=3.5.1
rich>=13.7.0
loguru>=0.7.2
nornir>=3.3.0
nornir-napalm>=0.4.0
napalm>=4.1.0